pyproject.toml
[tool.poetry]
name = "abr"
version = "0.1.0"
description = "A package to search for ABNs"
authors = ["Braidon Whatley <braidonw@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.15.0"
beautifulsoup4 = "^4.9.3"
lxml = "^4.6.2"
tqdm = "^4.58.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
